refactor: cypress 테스트 코드 리팩토링 및 파일 분리
